refactor(sql-editor): extract snippet-identity, mount, prettify, title hooks
Pulls four cohesive concerns out of the SQLEditor composition root into co-located hooks: useSnippetIdentity (id/name/loading derivation), useEditorMount (onMount + mount counter), usePrettifyQuery, and useSnippetTitleGenerator (title mutation + setAiTitle). Behavior-preserving; the characterization suite stays green. PR 3b of the SQLEditor decomposition stack.
